+++ tests/core/executor/test_MachineExecutor.py
@@ -3,11 +3,15 @@
 from pathlib import Path
 from unittest import TestCase
 
+from mockito import mock, unstub, ANY, verify, spy
+
 from vzmi.ychaos.core.exceptions.executor_errors import (
     YChaosTargetConfigConditionFailedError,
 )
-from vzmi.ychaos.core.executor.MachineTargetExecutor import MachineTargetExecutor
-from vzmi.ychaos.testplan.attack import AttackConfig, TargetType
+from vzmi.ychaos.core.executor.MachineTargetExecutor import (
+    MachineTargetExecutor,
+    YChaosAnsibleResultCallback,
+)
 from vzmi.ychaos.testplan.schema import TestPlan
 
 
@@ -32,3 +36,78 @@ class TestMachineExecutor(TestCase):
             self.testplans_directory.joinpath("valid/testplan2.yaml")
         )
         executor = MachineTargetExecutor(mock_valid_testplan)
+
+    def test_machine_executor_prepare(self):
+        mock_valid_testplan = TestPlan.load_file(
+            self.testplans_directory.joinpath("valid/testplan2.yaml")
+        )
+        executor = MachineTargetExecutor(mock_valid_testplan)
+        executor.prepare()
+
+        self.assertListEqual(
+            sorted(executor.ansible_context.inventory.hosts),
+            ["mockhost01.ychaos.yahoo.com", "mockhost02.ychaos.yahoo.com"],
+        )
+        self.assertEqual(executor.ansible_context.play_source["connection"], "ssh")
+        self.assertEqual(executor.ansible_context.play_source["strategy"], "free")
+        self.assertTrue(
+            executor.ansible_context.play_source["hosts"]
+            in ["mockhost01.ychaos.yahoo.com", "mockhost02.ychaos.yahoo.com"]
+        )
+
+    def test_ychaos_ansible_callback(self):
+        callback = YChaosAnsibleResultCallback(
+            hooks=dict(
+                on_target_passed=[lambda: self.assertTrue(True)],
+                on_target_failed=[lambda: self.assertTrue(True)],
+                on_target_unreachable=[lambda: self.assertTrue(True)],
+            )
+        )
+
+        mock_host_object = mock(dict(get_name=lambda: "mockhost01.ychaos.yahoo.com"))
+
+        mock_result = mock(dict(_host=mock_host_object))
+
+        callback.v2_runner_on_ok(mock_result)
+        self.assertEqual(len(callback.hosts_passed), 1)
+
+        callback.v2_runner_on_failed(mock_result)
+        self.assertEqual(len(callback.hosts_failed), 1)
+
+        callback.v2_runner_on_unreachable(mock_result)
+        self.assertEqual(len(callback.hosts_unreachable), 1)
+
+    def test_machine_executor_execute(self):
+        mock_valid_testplan = TestPlan.load_file(
+            self.testplans_directory.joinpath("valid/testplan2.yaml")
+        )
+        executor = MachineTargetExecutor(mock_valid_testplan)
+
+        class MockHookForTargetUnreachable:
+            # Always gets called for UT
+
+            def __init__(self):
+                self.test_value = False
+
+            def __call__(self, *args, **kwargs):
+                # Toggle the test value and assert this method is called
+                self.test_value = True
+
+        mock_hook_target_unreachable = MockHookForTargetUnreachable()
+
+        class MockHookForTargetPassed:
+            def __init__(self):
+                self.test_value = False
+
+            def __call__(self, *args, **kwargs):
+                # Should never get called
+                assert True is False
+
+        executor.register_hook("on_target_unreachable", mock_hook_target_unreachable)
+        executor.register_hook("on_target_passed", MockHookForTargetPassed())
+
+        executor.execute()
+        self.assertTrue(mock_hook_target_unreachable.test_value)
+
+    def tearDown(self) -> None:
+        unstub()